chore(clerk-js): Move useInView inside `/hooks`
+ Add default values in useLoadingStatus
