+++ packages/clerk-js/src/ui/hooks/index.ts
@@ -4,6 +4,7 @@ export * from './useWindowEventListener';
 export * from './useMagicLink';
 export * from './useClipboard';
 export * from './useEnabledThirdPartyProviders';
+export * from './useInView';
 export * from './useLoadingStatus';
 export * from './usePassword';
 export * from './usePasswordComplexity';
